% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{cal_abund}
\alias{cal_abund}
\title{Calculate the taxonomic abundance at each taxonomic ranks.}
\usage{
cal_abund()
}
\value{
taxa_summary in object.
}
\description{
Calculate the taxonomic abundance at each taxonomic ranks.
}
\examples{
dataset$cal_abund()
str(dataset$taxa_summary)
}
