% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{cal_alphadiv}
\alias{cal_alphadiv}
\title{Calculate alpha diversity in microtable Class.}
\usage{
cal_alphadiv(measures = NULL, PD = FALSE)
}
\arguments{
\item{measures}{one or more indexes from "Observed", "Coverage", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher", "PD"; default NULL, using all those measures.}

\item{PD}{TRUE or FALSE, whether phylogenetic tree should be calculated, default FALSE.}
}
\value{
alpha_diversity in object.
}
\description{
Calculate alpha diversity in microtable Class.
}
\examples{
dataset$cal_alphadiv(measures = NULL, PD = FALSE)
class(dataset$alpha_diversity)
}
