% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{cal_betadiv}
\alias{cal_betadiv}
\title{Calculate beta diversity in microtable Class.}
\usage{
cal_betadiv(unifrac = FALSE)
}
\arguments{
\item{unifrac}{TRUE or FALSE, whether unifrac index should be calculated, default FALSE.}
}
\value{
beta_diversity in object.
}
\description{
Calculate beta diversity in microtable Class.
}
\examples{
dataset$cal_betadiv(unifrac = FALSE)
class(dataset$beta_diversity)
}
