% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{cal_betamntd}
\alias{cal_betamntd}
\title{Calculate betaMNTD.}
\usage{
cal_betamntd(abundance.weighted = FALSE, exclude.conspecifics = FALSE)
}
\arguments{
\item{abundance.weighted}{default FALSE; whether use weighted abundance}

\item{exclude.conspecifics}{default FALSE; see \code{\link{comdistnt}} in picante package.}
}
\value{
res_betamntd in object.
}
\description{
This code runs faster than \code{\link{comdistnt}} in picante package.
}
\examples{
dataset$cal_betamntd(abundance.weighted=FALSE)
}
