% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{cal_betampd}
\alias{cal_betampd}
\title{Calculate betaMPD.}
\usage{
cal_betampd(abundance.weighted = FALSE)
}
\arguments{
\item{abundance.weighted}{default FALSE; whether use weighted abundance}
}
\value{
res_betampd in object.
}
\description{
This code runs faster than \code{\link{comdist}} in picante package.
}
\examples{
t1$cal_betampd(abundance.weighted=FALSE)
}
