% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{cal_biogeo}
\alias{cal_biogeo}
\title{Predict functional potential using FAPROTAX.}
\usage{
cal_biogeo(code_path = "./FAPROTAX_1.2.1", keep_tem = TRUE)
}
\arguments{
\item{code_path}{default "./FAPROTAX_1.2.1"; the code folder, download from http://www.loucalab.com/archive/FAPROTAX/lib/php/index.php?section=Download.}

\item{keep_tem}{default FALSE; whether keep the intermediate file, that is, the otu_table_for_FAPROTAX.txt in local place.}
}
\value{
res_biogeo in object.
}
\description{
Predict functional potential using FAPROTAX.
}
\examples{
dataset$cal_biogeo(code_path = "./FAPROTAX_1.2.1")
}
