% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_env.R
\name{cal_cor}
\alias{cal_cor}
\title{Correlations between abundance data table and physicochemical data table.}
\usage{
cal_cor(
  use_data = c("Genus", "all", "other")[1],
  select_env_data = NULL,
  cor_method = c("pearson", "spearman", "kendall")[1],
  p_adjust_method = "fdr",
  p_adjust_type = c("Env", "Taxa", "Type")[1],
  add_abund_table = NULL,
  by_group = NULL,
  other_taxa = NULL,
  group_use = NULL,
  group_select = NULL,
  taxa_name_full = TRUE
)
}
\arguments{
\item{use_data}{default "Genus"; one of c("Genus", "all", "other"); Genus: genus abundance, all: all taxa abundance, other: provide additional data with other_taxa parameter.}

\item{select_env_data}{default NULL; numeric or character vector to select columns in env_data; if not provided, automatically select the columns with numeric attributes.}

\item{p_adjust_method}{default "fdr"; p.adjust method.}

\item{p_adjust_type}{default "Env"; one of c("Type", "Taxa", "Env"); p.adjust type; Env: physicochemical data; Taxa: taxa data; Type: group used.}

\item{add_abund_table}{default NULL; additional data table to be used. Samples must be rows.}

\item{by_group}{default NULL; numeric or character vector to select one column in env_data; correlations separately for groups.}

\item{other_taxa}{default NULL; provide additional taxa, see use_data parameter.}

\item{group_use}{default NULL; numeric or character vector to select one column in env_data for selecting samples; together with group_select.}

\item{group_select}{default NULL; the group name used; will retain samples within the group.}

\item{taxa_name_full}{default TRUE; Whether retain the complete taxonomic name of taxa.}

\item{method}{default "pearson"; one of c("pearson", "spearman", "kendall"); correlation method.}
}
\value{
res_cor in object.
}
\description{
Calculating any correlation between two columns from two tables.
}
\examples{
t2 <- trans_diff$new(dataset = dataset, method = "rf", group = "Group", rf_taxa_level = "Genus")
t1 <- trans_env$new(dataset = dataset, add_data = env_data[, 4:11])
t1$cal_cor(use_data = "other", p_adjust_method = "fdr", other_taxa = t2$res_rf$Taxa[1:40])
}
