% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_alpha.R
\name{cal_diff}
\alias{cal_diff}
\title{Test the difference of alpha diveristy across groups.}
\usage{
cal_diff(method = c("KW", "anova")[1], measures = NULL)
}
\arguments{
\item{method}{default "KW"; "KW" or "anova"; KW rank sum test or anova for the testing.}

\item{measures}{default NULL; a vector, if null, all indexes will be calculated.}
}
\value{
res_alpha_diff in object.
}
\description{
Test the difference of alpha diveristy across groups. If use anova, require agricolae package.
}
\examples{
dataset$cal_diff(method = "KW")
dataset$cal_diff(method = "anova")
}
