% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_beta.R
\name{cal_group_distance}
\alias{cal_group_distance}
\title{Tranform sample distances within groups or between groups.}
\usage{
cal_group_distance(within_group = TRUE)
}
\arguments{
\item{within_group}{default TRUE; whether transform sample distance within groups, if false, transform sample distance between any two groups.}
}
\value{
group_distance in object.
}
\description{
Tranform sample distances within groups or between groups.
}
\examples{
t1$cal_group_distance()
}
