% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_beta.R
\name{cal_manova}
\alias{cal_manova}
\title{Calculate perMANOVA.}
\usage{
cal_manova(
  cal_manova_all = FALSE,
  cal_manova_paired = FALSE,
  cal_manova_set = NULL,
  permutations = 999
)
}
\arguments{
\item{cal_manova_all}{default FALSE; whether manova is used for all data.}

\item{cal_manova_paired}{default FALSE; whether manova is used for all the paired groups.}

\item{cal_manova_set}{default NULL; specified group set for manova, see \code{\link{adonis}}.}

\item{permutations}{default 999; see permutations in \code{\link{adonis}}.}
}
\value{
res_manova in object.
}
\description{
Calculate perMANOVA.
}
\examples{
t1$cal_manova(cal_manova_all = TRUE)
}
