% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_env.R
\name{cal_mantel}
\alias{cal_mantel}
\title{Mantel test between beta diversity matrix and physicochemical data.}
\usage{
cal_mantel(
  select_env_data = NULL,
  partial_mantel = FALSE,
  add_matrix = NULL,
  use_measure = NULL,
  method = "pearson",
  ...
)
}
\arguments{
\item{select_env_data}{default NULL; numeric or character vector to select columns in env_data; if not provided, automatically select the columns with numeric attributes.}

\item{partial_mantel}{default FALSE; whether use partial mantel test.}

\item{add_matrix}{default NULL; additional distance matrix provided, if you donot want to use the beta diversity matrix in the dataset.}

\item{use_measure}{default NULL; name of beta diversity matrix. If necessary and not provided, use the first beta diversity matrix.}

\item{method}{default "pearson"; one of c("pearson", "spearman", "kendall"); correlation method.}
}
\value{
res_mantel in object.
}
\description{
Mantel test between beta diversity matrix and physicochemical data.
}
\examples{
t1$cal_mantel(use_measure = "bray")
}
