% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{cal_network}
\alias{cal_network}
\title{Calculate network.}
\usage{
cal_network(
  network_method = c("COR", "PGM")[1],
  p_thres = 0.01,
  COR_weight = TRUE,
  COR_p_adjust = "fdr",
  COR_cut = 0.6,
  COR_optimization = FALSE,
  COR_low_threshold = 0.4,
  PGM_meta_data = FALSE,
  PGM_sensitive = "true",
  PGM_heterogeneous = "true",
  with_module = TRUE,
  add_taxa_name = "Phylum",
  usename_rawtaxa_when_taxalevel_notOTU = FALSE
)
}
\arguments{
\item{network_method}{default "COR"; "COR" or "PGM"; COR: correlation based method; PGM: Probabilistic Graphical Models based method; doi: 10.1016/j.cels.2019.08.002}

\item{p_thres}{default .01; the p value threshold.}

\item{COR_weight}{default TRUE; whether use correlation coefficient as the weight of edges.}

\item{COR_p_adjust}{default "fdr"; p.adjust method.}

\item{COR_cut}{default .6; correlation coefficient threshold.}

\item{COR_optimization}{default FALSE; whether use random matrix theory to optimize the choice of correlation coefficient, see https://doi.org/10.1186/1471-2105-13-113}

\item{COR_low_threshold}{default .4; the lowest correlation coefficient allowed.}

\item{PGM_meta_data}{default FALSE; whether use env data for the optimization, If TRUE, will automatically find the env_data in the object.}

\item{PGM_sensitive}{default "true"; whether use sensitive type in the PGM model.}

\item{PGM_heterogeneous}{default "true"; whether use heterogeneous type in the PGM model.}

\item{with_module}{default TRUE; whether calculate modules.}

\item{add_taxa_name}{default "Phylum"; add taxonomic rank name to the result.}

\item{usename_rawtaxa_when_taxalevel_notOTU}{default FALSE; whether replace the name of nodes using the taxonomic information.}
}
\value{
res_network in object.
}
\description{
Calculate network either based on the correlation method or based on the Probabilistic Graphical Models in julia FlashWeave.
}
\examples{
dataset$cal_network(p_thres = 0.01, COR_cut = 0.6)
}
