% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{cal_network_attr}
\alias{cal_network_attr}
\title{Calculate network properties.}
\usage{
cal_network_attr()
}
\value{
res_network_attr in object.
}
\description{
Calculate network properties.
}
\examples{
t1$cal_network_attr()
}
