% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{cal_node_type}
\alias{cal_node_type}
\title{Calculate node properties.}
\usage{
cal_node_type()
}
\value{
res_node_type in object.
}
\description{
Calculate node properties.
}
\examples{
t1$cal_node_type()
}
