% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{cal_process}
\alias{cal_process}
\title{Infer the processes according to ses.betaMNTD ses.betaMPD and rcbray.}
\usage{
cal_process(use_betamntd = TRUE)
}
\arguments{
\item{use_betamntd}{default TRUE; whether use ses.betaMNTD; if false, use ses.betaMPD.}
}
\value{
res_rcbray in object.
}
\description{
Infer the processes according to ses.betaMNTD ses.betaMPD and rcbray.
}
\examples{
t1$cal_process(use_betamntd = TRUE)
}
