% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{cal_prok_biogeo}
\alias{cal_prok_biogeo}
\title{Confirm species traits.}
\usage{
cal_prok_biogeo()
}
\value{
res_prok_biogeo in object.
}
\description{
Confirm traits of each OTU by matching the taxonomic assignments to the FAPROTAX database.
}
\examples{
t1$cal_prok_biogeo()
}
