% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{cal_prok_biogeo_perc}
\alias{cal_prok_biogeo_perc}
\title{Calculating the percentages of species with specific trait in communities or modules.}
\usage{
cal_prok_biogeo_perc(use_community = TRUE, node_type_table = NULL)
}
\arguments{
\item{use_community}{default TRUE; whether calculate community; if FALSE, use module.}

\item{node_type_table}{default NULL; If use_community FALSE; provide the node_type_table with the module information, such as the result of \code{\link{cal_node_type}}.}
}
\value{
res_prok_biogeo_perc in object.
}
\description{
The percentages of the OTUs with specific trait can reflect the potential of the corresponding function in the community or the module in the network.
}
\examples{
t1$cal_prok_biogeo_perc(use_community = TRUE)
}
