% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{cal_rcbray}
\alias{cal_rcbray}
\title{Calculate rcbray.}
\usage{
cal_rcbray(runs = 1000)
}
\arguments{
\item{runs}{default 1000; simulation runs.}
}
\value{
res_rcbray in object.
}
\description{
Calculate rcbray.
}
\examples{
dataset$cal_rcbray(runs=1000)
}
