% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_env.R
\name{cal_rda}
\alias{cal_rda}
\title{calculate RDA.}
\usage{
cal_rda(
  use_dbrda = TRUE,
  add_matrix = NULL,
  use_measure = NULL,
  feature_sel = FALSE,
  taxa_level = NULL,
  taxa_filter_thres = NULL
)
}
\arguments{
\item{use_dbrda}{default TRUE; whether use db-RDA, if FALSE, use RDA.}

\item{add_matrix}{default NULL; additional distance matrix provided, if you donot want to use the beta diversity matrix in the dataset.}

\item{use_measure}{default NULL; name of beta diversity matrix. If necessary and not provided, use the first beta diversity matrix.}

\item{feature_sel}{default FALSE; whether perform the feature selection.}

\item{taxa_level}{default NULL; If use RDA, provide the taxonomic rank.}

\item{taxa_filter_thres}{default NULL; If want to filter taxa, provide the relative abundance threshold.}
}
\value{
res_rda in object.
}
\description{
calculate RDA.
}
\examples{
dataset$cal_rda(use_dbrda = TRUE, use_measure = "bray")
}
