% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{cal_ses_betamntd}
\alias{cal_ses_betamntd}
\title{Calculate ses_betaMNTD.}
\usage{
cal_ses_betamntd(
  runs = 1000,
  abundance.weighted = FALSE,
  exclude.conspecifics = FALSE
)
}
\arguments{
\item{runs}{default 1000; simulation runs.}

\item{abundance.weighted}{default FALSE; whether use weighted abundance}

\item{exclude.conspecifics}{default FALSE; see \code{\link{comdistnt}} in picante package.}
}
\value{
res_ses_betamntd in object.
}
\description{
Calculate ses_betaMNTD.
}
\examples{
dataset$cal_ses_betamntd(runs=1000, abundance.weighted = FALSE, exclude.conspecifics = FALSE)
}
