% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{cal_ses_betampd}
\alias{cal_ses_betampd}
\title{Calculate ses_betaMPD.}
\usage{
cal_ses_betampd(runs = 1000, abundance.weighted = FALSE)
}
\arguments{
\item{runs}{default 1000; simulation runs.}

\item{abundance.weighted}{default FALSE; whether use weighted abundance}
}
\value{
res_ses_betampd in object.
}
\description{
Calculate ses_betaMPD.
}
\examples{
t1$cal_ses_betampd(runs=1000, abundance.weighted = FALSE)
}
