% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{cal_spe_func}
\alias{cal_spe_func}
\title{Confirm species traits.}
\usage{
cal_spe_func()
}
\value{
otu_func_table in object.
}
\description{
Confirm traits of each OTU by matching the taxonomic assignments to the FAPROTAX database.
}
\examples{
t1$cal_spe_func()
}
