% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{cal_tax4fun_func}
\alias{cal_tax4fun_func}
\title{Predict functional potential using tax4fun.}
\usage{
cal_tax4fun_func(keep_tem = FALSE, folderReferenceData = NULL)
}
\arguments{
\item{keep_tem}{default FALSE; whether keep the intermediate file, that is, the otu table in local place.}

\item{folderReferenceData}{default NULL; the folder, see http://tax4fun.gobics.de/ and \code{\link{Tax4Fun}} function in Tax4Fun package.}
}
\value{
tax4fun_KO and tax4fun_path in object.
}
\description{
Predict functional potential using tax4fun.
}
\examples{
dataset$cal_tax4fun_func(folderReferenceData = "./SILVA123")
}
