% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{clr_matrix}
\alias{clr_matrix}
\title{Centered log ratio tranformation for the matrix}
\usage{
clr_matrix(data, mar = 2, base = exp(1), tol = .Machine$double.eps)
}
\arguments{
\item{data}{a matrix}

\item{mar}{Margin, default 2, 1 = rows and 2 = columns of data.}

\item{base}{logarithmic base, default exp(1)}

\item{tol}{smallest representation of a residual of unity subtraction, default .Machine$double.eps}
}
\value{
transformed matrix
}
\description{
Centered log ratio tranformation for the matrix
}
