% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{color_palette_20}
\alias{color_palette_20}
\title{A color palette for 20 elements.}
\format{
An object of class \code{character} of length 20.
}
\usage{
color_palette_20
}
\description{
This is one palette option for users who have <= 20 elements to plot.
The palletes of RColorBrewer package provide at most 12 discrete colors, such as "Set3" and "Paired".
This palette is adapted from D3.js library and can be used as a supplement.
}
\keyword{datasets}
