% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert a symmetric correlation matrix to a covariance matrix
given the standard deviation}
\usage{
cor2cov(cor, sds)
}
\arguments{
\item{cor}{a symmetric correlation matrix}

\item{sds}{standard deviations of the resulting covariance.}
}
\value{
Covariance matrix of sample dimension as cor
}
\description{
Convert a symmetric correlation matrix to a covariance matrix
given the standard deviation
}
