% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env_data_16S}
\alias{env_data_16S}
\title{The environmental factors for the 16S dataset in the microeco package}
\usage{
data(env_data_16S)
}
\description{
The environmental factors for the 16S dataset in the microeco package
}
\keyword{data.frame}
