% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{filter_pollution}
\alias{filter_pollution}
\title{Filter the taxa considered as pollution.}
\usage{
filter_pollution(taxa = c("mitochondria", "chloroplast"))
}
\arguments{
\item{taxa}{default: c("mitochondria", "chloroplast"); filter mitochondria and chloroplast, or others as needed.}
}
\value{
None
}
\description{
This operation will remove any line of the tax_table containing any the word in taxa parameter regardless of word case.
}
\examples{
dataset$filter_pollution(taxa = c("mitochondria", "chloroplast"))
}
