% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fungi_func}
\alias{fungi_func}
\title{The fungi function database in the microeco package}
\usage{
data(fungi_func)
}
\description{
The fungi function database in the microeco package
}
\keyword{data.frame}
