% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fungi_func_FungalTraits}
\alias{fungi_func_FungalTraits}
\title{The FungalTraits database for fungi trait identification in the microeco package}
\usage{
data(fungi_func_FungalTraits)
}
\description{
The FungalTraits database for fungi trait identification in the microeco package
}
\keyword{data.frame}
