% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ko_map}
\alias{ko_map}
\title{The KEGG pathway annotation database in the microeco package}
\usage{
data(ko_map)
}
\description{
The KEGG pathway annotation database in the microeco package
}
\keyword{data.frame}
