% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{meco2phyloseq}
\alias{meco2phyloseq}
\title{Transform microtable object in microeco package to the phyloseq object in phyloseq package.}
\usage{
meco2phyloseq(dataset)
}
\arguments{
\item{dataset}{a microtable object.}
}
\value{
phyloseq object.
}
\description{
Transform microtable object in microeco package to the phyloseq object in phyloseq package.
}
\examples{
\donttest{
data("dataset")
meco2phyloseq(dataset)
}
}
