% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples according to specific group to generate a new microtable.}
\usage{
merge_samples(use_group)
}
\arguments{
\item{use_group}{the specific group in sample_table.}
}
\value{
a new created merged \code{\link{microtable}}.
}
\description{
Merge samples according to specific group to generate a new microtable.
}
\examples{
dataset$merge_samples(use_group = "Group")
}
