% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{merge_taxa}
\alias{merge_taxa}
\title{Merge taxa according to specific taxonomic rank to generate a new microtable.}
\usage{
merge_taxa(taxa = "Genus")
}
\arguments{
\item{taxa}{the specific rank in tax_table.}
}
\value{
a new created merged \code{\link{microtable}}.
}
\description{
Merge taxa according to specific taxonomic rank to generate a new microtable.
}
\examples{
dataset$merge_taxa(taxa = "Genus")
}
