% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm_to_total}
\alias{norm_to_total}
\title{Total sum normalization of a (presumably positive) data vector}
\usage{
norm_to_total(x)
}
\description{
Total sum normalization of a (presumably positive) data vector
}
