% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_table_16S}
\alias{otu_table_16S}
\title{The OTU table of the 16S dataset in the microeco package}
\usage{
data(otu_table_16S)
}
\description{
The OTU table of the 16S dataset in the microeco package
}
\keyword{data.frame}
