% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_table_ITS}
\alias{otu_table_ITS}
\title{The OTU table of the ITS example data}
\format{
data.frame
}
\usage{
data(otu_table_ITS)
}
\description{
The OTU table of the ITS example data
}
\keyword{datasets}
