% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{phyloseq2meco}
\alias{phyloseq2meco}
\title{Transform the phyloseq object in phyloseq package to microtable object in microeco package.}
\usage{
phyloseq2meco(physeq)
}
\arguments{
\item{physeq}{a phyloseq object.}
}
\value{
microtable object.
}
\description{
Transform the phyloseq object in phyloseq package to microtable object in microeco package.
}
\examples{
\donttest{
library(phyloseq)
data("GlobalPatterns")
phyloseq2meco(GlobalPatterns)
}
}
