% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_alpha.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Plotting the alpha diveristy.}
\usage{
plot_alpha(
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  measure = "Shannon",
  group = NULL,
  add_letter = FALSE,
  pair_compare = FALSE,
  pair_compare_filter = "",
  pair_compare_method = "wilcox.test",
  map_signif_level = TRUE,
  xtext_type = NULL,
  xtext_size = 10,
  ytitle_size = 17,
  base_font = NULL,
  ...
)
}
\arguments{
\item{color_values}{colors used for presentation.}

\item{measure}{default Shannon; alpha diveristy that will be plotted.}

\item{group}{default NULL; group name used for the plot.}

\item{add_letter}{default FALSE; If TRUE, the letters of duncan test will be added in the plot.}

\item{pair_compare}{default FALSE; whether do paired comparisons.}

\item{pair_compare_filter}{default ""; groups that will be removed.}

\item{pair_compare_method}{default wilcox.test; wilcox.test, kruskal.test, t.test or anova.}

\item{map_signif_level}{default TRUE; whether indicate the significance level.}

\item{xtext_type}{default NULL; number used to make x axis text generate angle.}

\item{xtext_size}{default 10, x axis text size.}

\item{ytitle_size}{default 17, y axis title size.}

\item{base_font}{default NULL, font in the plot.}

\item{...}{parameters pass to \code{\link{ggpubr::ggboxplot}}.}
}
\value{
ggplot.
}
\description{
Plotting the alpha diveristy.
}
\examples{
dataset$plot_alpha(color_values = RColorBrewer::brewer.pal(12, "Paired"), measure = "Shannon", group = "Group", pair_compare = TRUE)
}
