% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_abund.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot the bar plot with the object of trans_abund Class.}
\format{
\code{\link{trans_abund}} object.
}
\usage{
plot_bar(
  use_colors = RColorBrewer::brewer.pal(12, "Paired"),
  bar_type = "full",
  others_color = "grey90",
  facet = NULL,
  barwidth = NULL,
  facet_color = "grey95",
  strip_text = 11,
  legend_text_italic = FALSE,
  xtext_type_hor = TRUE,
  xtext_size = 10,
  xtext_keep = TRUE,
  xtitle_keep = TRUE,
  ytitle_size = 17,
  base_font = NULL,
  ylab_title = NULL
)
}
\arguments{
\item{bar_type}{default "full"; "full" or "notfull"; if full, the total abundance sum to 1 or 100 percentage.}

\item{others_color}{default "grey90"; the color for others.}

\item{facet}{default NULL; if using facet, providing the group name.}

\item{barwidth}{default NULL; bar width, see width in \code{\link{geom_bar}}.}

\item{facet_color}{default "grey95"; facet background color.}

\item{strip_text}{default 11; facet text size.}

\item{legend_text_italic}{default FALSE; whether use italic in legend.}

\item{xtext_type_hor}{default TRUE; x axis text horizontal, if FALSE; text slant.}

\item{xtext_size}{default 10; x axis text size.}

\item{xtext_keep}{default TRUE; whether retain x text.}

\item{xtitle_keep}{default TRUE; whether retain x title.}

\item{ytitle_size}{default 17; y axis title size.}

\item{base_font}{default NULL; font in the plot.}

\item{ylab_title}{default NULL; y axis title.}

\item{bar_colors}{default RColorBrewer::brewer.pal(12, "Paired"); providing the plotting colors.}
}
\value{
ggplot2 plot.
}
\description{
Plot the bar plot with the object of trans_abund Class.
}
\examples{
t1$plot_bar(bar_type = "full", others_color = "grey90")
}
