% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_abund.R
\name{plot_box}
\alias{plot_box}
\title{Plot the box plot with the object of trans_abund Class.}
\format{
\code{\link{trans_abund}} object.
}
\usage{
plot_box(
  use_colors = RColorBrewer::brewer.pal(12, "Paired"),
  group = NULL,
  show_point = FALSE,
  point_color = "black",
  point_size = 3,
  point_alpha = 0.3,
  plot_flip = FALSE,
  boxfill = TRUE,
  middlecolor = "grey95",
  middlesize = 1,
  xtext_type_hor = TRUE,
  xtext_size = 10,
  xtext_keep = TRUE,
  xtitle_keep = TRUE,
  ytitle_size = 17,
  base_font = NULL,
  ...
)
}
\arguments{
\item{use_colors}{default RColorBrewer::brewer.pal(12, "Paired"); providing the plotting colors.}

\item{group}{default NULL; sample table column name.}

\item{show_point}{default FALSE; whether show points in plot.}

\item{point_color}{default "black"; If show_point TRUE; use the color}

\item{point_size}{default 3; If show_point TRUE; use the size}

\item{point_alpha}{default .3; If show_point TRUE; use the transparency.}

\item{plot_flip}{default FALSE; Whether rotate plot.}

\item{boxfill}{default TRUE; Whether fill the box.}

\item{middlecolor}{default "grey95"; The middle line color.}

\item{middlesize}{default 1; The middle line size.}

\item{xtext_type_hor}{default TRUE; x axis text horizontal, if FALSE; text slant.}

\item{xtext_size}{default 10; x axis text size.}

\item{xtext_keep}{default TRUE; whether retain x text.}

\item{xtitle_keep}{default TRUE; whether retain x title.}

\item{ytitle_size}{default 17; y axis title size.}

\item{base_font}{default NULL; font in the plot.}

\item{...}{parameters pass to \code{\link{geom_boxplot}}.}
}
\value{
ggplot2 plot.
}
\description{
Plot the box plot with the object of trans_abund Class.
}
\examples{
t1$plot_box(group = "Group", show_point = FALSE)
}
