% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_beta.R
\name{plot_clustering}
\alias{plot_clustering}
\title{Plotting clustering result.}
\usage{
plot_clustering(
  use_colors = RColorBrewer::brewer.pal(8, "Dark2"),
  measure = NULL,
  group = NULL,
  replace_name = NULL
)
}
\arguments{
\item{use_colors}{colors for presentation.}

\item{measure}{default NULL; beta diversity index; suggest NULL, use the measure when creating object}

\item{group}{default NULL; if provided, use this group to assign color.}

\item{replace_name}{default NULL; if provided, use this as label.}

\item{plot_group_order}{default NULL; a vector used to order the groups in the plot.}
}
\value{
ggplot.
}
\description{
Plotting clustering result. Require ggdendro package.
}
\examples{
t1$plot_clustering(group = "Group", replace_name = c("Saline", "Type"))
}
