% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_env.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot correlations heatmap.}
\usage{
plot_corr(
  color_vector = c("#00008B", "#102D9B", "#215AAC", "#3288BD", "#66C2A5", "#E6F598",
    "#FFFFBF", "#FED690", "#FDAE61", "#F46D43", "#D53E4F"),
  pheatmap = FALSE,
  ylab_type_italic = FALSE,
  keep_full_name = FALSE,
  keep_prefix = TRUE,
  plot_x_size = 9,
  mylabels_x = NULL,
  font_family = NULL
)
}
\arguments{
\item{color_vector}{color pallete.}

\item{pheatmap}{default FALSE; whether use heatmap with clustering plot.}

\item{ylab_type_italic}{default FALSE; whether use italic type for y lab text.}

\item{keep_full_name}{default FALSE; whether use the complete taxonomic name.}

\item{keep_prefix}{default TRUE; whether retain the taxonomic prefix.}

\item{plot_x_size}{default 9; x axis text size.}

\item{mylabels_x}{default NULL; provide x axis text labels additionally; only available when pheatmap = TRUE.}

\item{font_family}{default NULL; font family used in ggplot2; only available when pheatmap = FALSE.}
}
\value{
plot.
}
\description{
Plot correlations heatmap.
}
\examples{
t1$plot_corr()
}
