% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_diff.R
\name{plot_diff_abund}
\alias{plot_diff_abund}
\title{Plotting the abundance of differencial taxa.}
\usage{
plot_diff_abund(
  method = NULL,
  only_abund_plot = TRUE,
  use_number = 1:10,
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  plot1_bar_color = "grey50",
  plot2_sig_color = "red",
  plot2_sig_size = 1.2,
  axis_text_y = 10,
  simplify_names = TRUE,
  keep_prefix = TRUE,
  group_order = NULL,
  plot2_barwidth = 0.9,
  add_significance = TRUE,
  use_se = TRUE
)
}
\arguments{
\item{method}{default NULL; "rf" or "lefse"; automatically check the method in the result.}

\item{only_abund_plot}{default TRUE; if true, return only abundance plot; if false, return both indicator plot and abundance plot}

\item{use_number}{default 1:10; vector, the taxa numbers used in the plot, 1:n.}

\item{color_values}{colors for presentation.}

\item{plot1_bar_color}{default "grey30"; the color for the plot 1.}

\item{plot2_sig_color}{default "red"; the color for the significance in plot 2.}

\item{plot2_sig_size}{default 1.5; the size for the significance in plot 2.}

\item{axis_text_y}{default 12; the size for the y axis text.}

\item{simplify_names}{default TRUE; whether use the simplified taxonomic name.}

\item{keep_prefix}{default TRUE; whether retain the taxonomic prefix.}

\item{group_order}{default NULL; a vector to order the legend in plot.}

\item{plot2_barwidth}{default .9; the bar width in plot 2.}

\item{add_significance}{default TRUE; whether add the significance asterisk; only available when only_abund_plot FALSE.}

\item{use_se}{default TRUE; whether use SE in plot 2, if FALSE, use SD.}
}
\value{
a list with two ggplot.
}
\description{
Plotting the abundance of differencial taxa.
}
\examples{
dataset$plot_diff_abund(use_number = 1:10)
}
