% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_beta.R
\name{plot_group_distance}
\alias{plot_group_distance}
\title{Plotting the distance between samples within or between groups.}
\usage{
plot_group_distance(
  plot_group_order = NULL,
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  distance_pair_stat = FALSE,
  pair_compare_filter = "",
  pair_compare_method = "wilcox.test",
  map_signif_level = TRUE,
  plot_distance_xtype = NULL
)
}
\arguments{
\item{plot_group_order}{default NULL; a vector used to order the groups in the plot.}

\item{color_values}{colors for presentation.}

\item{distance_pair_stat}{default FALSE; whether do the paired comparisions.}

\item{pair_compare_filter}{default ""; if provided, remove the matched groups.}

\item{pair_compare_method}{default wilcox.test; wilcox.test, kruskal.test, t.test or anova.}

\item{map_signif_level}{default TRUE; whether indicate the significance level.}

\item{plot_distance_xtype}{default NULL; number used to make x axis text generate angle.}
}
\value{
ggplot.
}
\description{
Plotting the distance between samples within or between groups.
}
\examples{
t1$plot_group_distance(distance_pair_stat = TRUE)
}
