% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_diff.R
\name{plot_lefse_bar}
\alias{plot_lefse_bar}
\title{Bar plot for LDA score.}
\usage{
plot_lefse_bar(
  use_number = 1:10,
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  LDA_score = NULL,
  simplify_names = TRUE,
  keep_prefix = TRUE,
  group_order = NULL,
  axis_text_y = 12,
  plot_vertical = TRUE,
  ...
)
}
\arguments{
\item{use_number}{default 1:10; vector, the taxa numbers used in the plot, 1:n.}

\item{color_values}{colors for presentation.}

\item{LDA_score}{default NULL; numeric value as the threshold, such as 2, limited with use_number.}

\item{simplify_names}{default TRUE; whether use the simplified taxonomic name.}

\item{keep_prefix}{default TRUE; whether retain the taxonomic prefix.}

\item{group_order}{default NULL; a vector to order the legend in plot.}

\item{axis_text_y}{default 12; the size for the y axis text.}

\item{plot_vertical}{default TRUE; whether use vertical bar plot or horizontal.}

\item{...}{parameters pass to \code{\link{geom_bar}}}
}
\value{
ggplot.
}
\description{
Bar plot for LDA score.
}
\examples{
t1$plot_lefse_bar(LDA_score = 2)
}
