% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_diff.R
\name{plot_lefse_cladogram}
\alias{plot_lefse_cladogram}
\title{Plot the cladogram for LEfSe result.}
\usage{
plot_lefse_cladogram(
  color = RColorBrewer::brewer.pal(8, "Dark2"),
  use_taxa_num = 200,
  filter_taxa = NULL,
  use_feature_num = NULL,
  clade_label_level = 4,
  select_show_labels = NULL,
  only_select_show = FALSE,
  sep = "|",
  branch_size = 0.2,
  alpha = 0.2,
  clade_label_size = 0.7,
  node_size_scale = 1,
  node_size_offset = 1,
  annotation_shape = 22,
  annotation_shape_size = 5
)
}
\arguments{
\item{color}{default RColorBrewer::brewer.pal(8, "Dark2"); color used in the plot.}

\item{use_taxa_num}{default 200; integer; The taxa number used in the background tree plot; select the taxa according to the mean abundance}

\item{filter_taxa}{default NULL; The mean relative abundance used to filter the taxa with low abundance}

\item{use_feature_num}{default NULL; integer; The feature number used in the plot; select the features according to the LDA score}

\item{clade_label_level}{default 4; the taxonomic level for marking the label with letters, root is the largest}

\item{select_show_labels}{default NULL; character vector; The features to show in the plot with full label names, not the letters}

\item{only_select_show}{default FALSE; whether only use the the select features in the parameter select_show_labels}

\item{sep}{default "|"; the seperate character in the taxonomic information}

\item{branch_size}{default 0.2; numberic, size of branch}

\item{alpha}{default 0.2; shading of the color}

\item{clade_label_size}{default 0.7; size for the clade label}

\item{node_size_scale}{default 1; scale for the node size}

\item{node_size_offset}{default 1; offset for the node size}

\item{annotation_shape}{default 22; shape used in the annotation legend}

\item{annotation_shape_size}{default 5; size used in the annotation legend}
}
\value{
ggplot.
}
\description{
Plot the cladogram for LEfSe result similar with the python version. Codes are modified from microbiomeMarker
}
\examples{
t1$plot_lefse_cladogram(use_taxa_num = 200, use_feature_num = 50, select_show_labels = NULL)
}
