% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_beta.R
\name{plot_ordination}
\alias{plot_ordination}
\title{Plotting the ordination result.}
\usage{
plot_ordination(
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  plot_color = NULL,
  plot_shape = NULL,
  plot_group_order = NULL,
  shape_values = c(16, 17, 7, 8, 15, 18, 11, 10, 12, 13, 9, 3, 4, 0, 1, 2, 14),
  plot_point_size = 3,
  plot_point_alpha = 0.9,
  plot_sample_label = NULL,
  plot_group_centroid = FALSE,
  plot_group = NULL,
  segment_alpha = 0.6,
  centroid_linetype = 3,
  plot_group_ellipse = FALSE,
  ellipse_level = 0.9,
  ellipse_alpha = 0.1,
  ellipse_type = "t"
)
}
\arguments{
\item{color_values}{colors for presentation.}

\item{plot_color}{default NULL; the sample group name used for color in plot.}

\item{plot_shape}{default NULL; the sample group name used for shape in plot.}

\item{plot_group_order}{default NULL; a vector used to order the groups in the legend of plot.}

\item{shape_values}{default c(16, 17, 7,  8, 15); a vector used in the shape type of plot, see ggplot2 tutorial.}

\item{plot_point_size}{default 3; point size in plot.}

\item{plot_point_alpha}{default .9; point transparency in plot.}

\item{plot_sample_label}{default NULL; the column name in sample table, if provided, show the point name in plot.}

\item{plot_group_centroid}{default FALSE; whether show the centroid in each group of plot.}

\item{plot_group}{default NULL; the column name in sample table, generally used with plot_group_centroid and plot_group_ellipse.}

\item{segment_alpha}{default .6; segment transparency in plot.}

\item{plot_group_ellipse}{default FALSE; whether show the confidence ellipse in each group of plot.}

\item{ellipse_level}{default .9; confidence level of ellipse.}

\item{ellipse_alpha}{default .1; color transparency in the ellipse.}

\item{ellipse_type}{default t; see type in \code{\link{stat_ellipse}}.}
}
\value{
ggplot.
}
\description{
Plotting the ordination result.
}
\examples{
dataset$plot_ordination(plot_color = "Group", plot_shape = "Group", plot_group_ellipse = TRUE)
}
