% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{plot_prok_biogeo_perc}
\alias{plot_prok_biogeo_perc}
\title{Plot the percentages of species with specific trait in communities or modules.}
\usage{
plot_prok_biogeo_perc(
  filter_func = NULL,
  group_list = NULL,
  group_list_default = FALSE,
  add_facet = TRUE,
  select_samples = NULL
)
}
\arguments{
\item{filter_func}{default NULL; a vector of function names.}

\item{group_list}{default NULL; a list with group names and the functions in the groups.}

\item{group_list_default}{default FALSE; whether use the default group list.}

\item{add_facet}{default TRUE; whether use facet in the plot.}

\item{select_samples}{default NULL; character vector, select partial samples to show}
}
\value{
ggplot2.
}
\description{
Plot the percentages of species with specific trait in communities or modules.
}
\examples{
t1$plot_prok_biogeo_perc(group_list_default = TRUE)
}
