% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_env.R
\name{plot_rda}
\alias{plot_rda}
\title{plot RDA result.}
\usage{
plot_rda(
  plot_color = NULL,
  plot_shape = NULL,
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  taxa_text_color = "firebrick1",
  taxa_text_type = "italic"
)
}
\arguments{
\item{plot_color}{default NULL; group used for color.}

\item{plot_shape}{default NULL; group used for shape.}

\item{color_values}{color pallete.}

\item{taxa_text_color}{default "firebrick1"; taxa text colors.}

\item{taxa_text_type}{default "italic"; taxa text style; better to use "italic" for Genus, use "normal" for others.}
}
\value{
ggplot object.
}
\description{
plot RDA result.
}
\examples{
t1$plot_rda(plot_color = "Group")
}
