% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{plot_spe_func_perc}
\alias{plot_spe_func_perc}
\title{Plot the percentages of species with specific trait in communities or modules.}
\usage{
plot_spe_func_perc(
  filter_func = NULL,
  group_list = NULL,
  group_list_default = FALSE
)
}
\arguments{
\item{filter_func}{default NULL; a vector of function names.}

\item{group_list}{default NULL; a list with group names and the functions in the groups.}

\item{group_list_default}{default FALSE; whether use the default group list.}
}
\value{
ggplot2.
}
\description{
Plot the percentages of species with specific trait in communities or modules.
}
\examples{
dataset$plot_spe_func_perc(group_list_default = TRUE)
}
