% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{plot_sum_links}
\alias{plot_sum_links}
\title{Plot the summed linkages among taxa.}
\usage{
plot_sum_links(plot_pos = TRUE, plot_num = NULL, color_values = NULL)
}
\arguments{
\item{plot_pos}{default TRUE; plot the summed positive or negative linkages.}

\item{plot_num}{default NULL; number of taxa presented in the plot.}

\item{color_values}{default NULL; If not provided, use default.}
}
\value{
chorddiag plot
}
\description{
Plot the summed linkages among taxa.
}
\examples{
t1$plot_sum_links(plot_pos = TRUE, plot_num = 10)
}
