% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{plot_taxa_roles}
\alias{plot_taxa_roles}
\title{Plot the classification and importance of nodes.}
\usage{
plot_taxa_roles(roles_colors = NULL, plot_module = FALSE)
}
\arguments{
\item{roles_colors}{default NULL; colors for each group.}

\item{plot_module}{default FALSE; whether plot the modules information.}
}
\value{
ggplot.
}
\description{
Plot the classification and importance of nodes.
}
\examples{
dataset$plot_taxa_roles()
}
