% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plot venn.}
\usage{
plot_venn(
  text_size = 4.5,
  text_name_size = 6,
  text_name_position = NULL,
  color_circle = RColorBrewer::brewer.pal(8, "Dark2"),
  alpha = 0.3,
  fill_color = TRUE,
  linesize = 1.1,
  petal_plot = FALSE,
  petal_color = "skyblue",
  petal_a = 4,
  petal_r = 1,
  petal_use_lim = c(-12, 12),
  petal_center_size = 40,
  petal_move_xy = 4,
  petal_move_k = 2.3,
  petal_move_k_count = 1.3,
  petal_text_move = 40
)
}
\arguments{
\item{text_size}{default 4.5; text size in plot}

\item{text_name_size}{default 6; name size in plot}

\item{text_name_position}{default NULL; name position in plot; use default, if NULL.}

\item{color_circle}{color pallete}

\item{alpha}{default .3; alpha for transparency}

\item{fill_color}{default TRUE; whether fill the area color}

\item{linesize}{default 1.1; cycle line size}

\item{petal_plot}{default FALSE; whether use petal plot.}

\item{petal_color}{default "skyblue"; color of the petal}

\item{petal_a}{default 4; the length of the ellipse}

\item{petal_r}{default 1; scaling up the size of the ellipse}

\item{petal_use_lim}{default c(-12, 12); the width of the plot}

\item{petal_center_size}{default 40; petal center circle size}

\item{petal_move_xy}{default 4; the distance of text to circle}

\item{petal_move_k}{default 2.3; the distance of title to circle}

\item{petal_move_k_count}{default 1.3; the distance of data text to circle}

\item{petal_text_move}{default 40; the distance between two data text}
}
\value{
ggplot.
}
\description{
Plot venn.
}
\examples{
dataset$plot_venn()
}
