% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prok_func}
\alias{prok_func}
\title{The prokaryotic trait database in the microeco package}
\usage{
data(prok_func)
}
\description{
The prokaryotic trait database in the microeco package
}
\keyword{data.frame}
