% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prok_func_FAPROTAX}
\alias{prok_func_FAPROTAX}
\title{Customized FAPROTAX trait database}
\format{
list
}
\usage{
data(prok_func_FAPROTAX)
}
\description{
Customized FAPROTAX trait database
}
\keyword{datasets}
