% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prok_func_NJC19_list}
\alias{prok_func_NJC19_list}
\title{The modified NJC19 database in the microeco package}
\usage{
data(prok_func_NJC19_list)
}
\description{
The modified NJC19 database in the microeco package
}
\keyword{data.frame}
