% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pval.sparccboot}
\alias{pval.sparccboot}
\title{SparCC p-vals}
\usage{
pval.sparccboot(x, sided = "both")
}
\arguments{
\item{x}{output from \code{sparccboot}}

\item{sided}{type of p-value to compute. Only two sided (sided="both") is implemented.}
}
\description{
Get empirical p-values from bootstrap SparCC output, modified based on https://github.com/zdk123/SpiecEasi
}
