% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{rarefy_samples}
\alias{rarefy_samples}
\title{Rarefy communities to make all samples have same species number.}
\usage{
rarefy_samples(sample.size = NULL, rngseed = 123, replace = TRUE)
}
\arguments{
\item{sample.size}{default:NULL; the required species number, If not provided, use minimum number of all samples.}

\item{rngseed}{random seed; default: 123.}

\item{replace}{default: TRUE; see \code{\link{sample}} for the random sampling.}
}
\value{
None; rarefied dataset.
}
\description{
Rarefy communities to make all samples have same species number.
}
\examples{
dataset$rarefy_samples(sample.size = min(dataset$sample_sums()), replace = TRUE)
}
