% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_fasta_16S}
\alias{rep_fasta_16S}
\title{The fasta file of 16S dataset used in tax4fun2 method.}
\usage{
data(rep_fasta_16S)
}
\description{
See the document of microtable class for more details. This file is with read.fasta function in seqinr package.
}
\keyword{list}
