% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{replace_name}
\alias{replace_name}
\title{Replace names in res_cor_p of trans_corr object.}
\usage{
replace_name()
}
\arguments{
\item{No}{parameters}
}
\value{
a new res_cor_p in trans_corr object.
}
\description{
Replace names in res_cor_p of trans_corr object.
}
\examples{
trans_corr$replace_name()
}
