% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_info_16S}
\alias{sample_info_16S}
\title{The sample information of 16S dataset in the microeco package}
\usage{
data(sample_info_16S)
}
\description{
The sample information of 16S dataset in the microeco package
}
\keyword{data.frame}
