% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_info_ITS}
\alias{sample_info_ITS}
\title{The sample information of ITS dataset in the microeco package}
\usage{
data(sample_info_ITS)
}
\description{
The sample information of ITS dataset in the microeco package
}
\keyword{data.frame}
