% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{sample_sums}
\alias{sample_sums}
\title{Sum the species number for each sample.}
\usage{
sample_sums()
}
\value{
species number of samples.
}
\description{
Sum the species number for each sample.
}
\examples{
dataset$sample_sums()
}
