% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{save_abund}
\alias{save_abund}
\title{Save taxonomic abundance to the computer.}
\usage{
save_abund(dirpath = "taxa_abund")
}
\arguments{
\item{dirpath}{default "taxa_abund"; directory name to save the taxonomic abundance files.}
}
\description{
Save taxonomic abundance to the computer.
}
\examples{
dataset$save_abund(dirpath = "taxa_abund")
}
