% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{save_alphadiv}
\alias{save_alphadiv}
\title{Save alpha diversity table to the computer.}
\usage{
save_alphadiv(dirpath = "alpha_diversity")
}
\arguments{
\item{dirpath}{default "alpha_diversity"; directory name to save the alpha_diversity.csv file.}
}
\description{
Save alpha diversity table to the computer.
}
\examples{
# save alpha diversity table in the alpha_diversity directory
dataset$save_alphadiv(dirpath = "alpha_diversity")
}
