% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{save_betadiv}
\alias{save_betadiv}
\title{Save beta diversity matrix to the computer.}
\usage{
save_betadiv(dirpath = "beta_diversity")
}
\arguments{
\item{dirpath}{default "beta_diversity"; directory name to save the beta diversity matrix files.}
}
\description{
Save beta diversity matrix to the computer.
}
\examples{
dataset$save_betadiv(dirpath = "beta_diversity")
}
