% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{save_network}
\alias{save_network}
\title{Save network.}
\usage{
save_network(filepath = "network.gexf")
}
\arguments{
\item{filepath}{default "network.gexf"; file path.}
}
\value{
None.
}
\description{
Save network.
}
\examples{
t1$save_network(filepath = "network.gexf")
}
