% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{show_spe_func}
\alias{show_spe_func}
\title{Show the basic information for one function.}
\usage{
show_spe_func(use_community = TRUE, node_type_table = NULL)
}
\arguments{
\item{use_func}{default NULL; the function name.}
}
\value{
None.
}
\description{
Show the basic information for one function.
}
\examples{
t1$show_spe_func(use_func = "methanotrophy")
}
