% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_metab}
\alias{soil_metab}
\title{Microtable object of non-targeted metabolome data from bulk soil and maize rhizosphere soil
in an agricultural field experiment with crop rotation and fertilization treatments (DOI: 10.1007/s11104-024-06847-9).}
\format{
R6 class object
}
\usage{
data(soil_metab)
}
\description{
For the explanation of each column of \code{sample_table}, please refer to the \code{soil_microb} data document.
In \code{tax_table}, 'HMDB_ID' is the number of HMDB database. 'KEGG_ID' is the number of KEGG database.
\itemize{
  \item sample_table: sample information table
  \item otu_table: metabolites abundance table
  \item tax_table: annotation table
}
}
\keyword{datasets}
