% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_microb}
\alias{soil_microb}
\title{Microtable object of prokaryotic sequencing data from soil and maize 
in an agricultural field experiment with crop rotation and fertilization treatments (DOI: 10.1007/s11104-024-06847-9).}
\format{
R6 class object
}
\usage{
data(soil_microb)
}
\description{
In the first column of \code{sample_table}, 'SampleID', is the sample name and same with the row names. 
The second column, 'Group', represents different treatment groups, namely the combinations of 'Cropping' and 'Fertilization', which are two experimental factors. 
'Cropping' denotes the experimental treatments including rotational (RC) and continuous (CC) cropping, 
and 'Fertilization' includes CK (no fertilizer control), NPK (inorganic fertilizer) and NPKS (NPK + straw amendment). 
'Compartment' means different sampling compartments (bulk soil, rhizosphere soil and root).
}
\details{
\itemize{
  \item sample_table: sample information table
  \item otu_table: ASV abundance table
  \item tax_table: taxonomic table
}
}
\keyword{datasets}
