% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{subset_network}
\alias{subset_network}
\title{Subset of the network.}
\usage{
subset_network(node = NULL, rm_single = TRUE)
}
\arguments{
\item{node}{default NULL; provide the names of the nodes that you want to use in the sub-network.}

\item{rm_single}{default TRUE; whether remove the nodes without any edge in the sub-network.}
}
\value{
network
}
\description{
Subset of the network.
}
\examples{
t1$subset_network(node = t1$res_node_type \%>\% .[.$module == "M1", ] \%>\% rownames, rm_single = TRUE)
# return a sub network that contains all nodes of module M1
}
