% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tax4Fun2_KEGG}
\alias{Tax4Fun2_KEGG}
\title{The KEGG data files used in the trans_func class}
\format{
list
}
\usage{
data(Tax4Fun2_KEGG)
}
\description{
The KEGG data files used in the trans_func class
}
\keyword{datasets}
