% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{taxa_sums}
\alias{taxa_sums}
\title{Sum the species number for each taxa.}
\usage{
taxa_sums()
}
\value{
species number of taxa.
}
\description{
Sum the species number for each taxa.
}
\examples{
dataset$taxa_sums()
}
