% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_table_ITS}
\alias{taxonomy_table_ITS}
\title{The taxonomic information of ITS dataset in the microeco package}
\usage{
data(taxonomy_table_ITS)
}
\description{
The taxonomic information of ITS dataset in the microeco package
}
\keyword{data.frame}
