% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtable.R
\name{tidy_dataset}
\alias{tidy_dataset}
\title{Tidy the object of microtable Class.}
\usage{
tidy_dataset(main_data = TRUE)
}
\arguments{
\item{main_data}{TRUE or FALSE, if TRUE, only basic files in microtable object is tidied, otherwise, all files, including taxa_summary, alpha_diversity and beta_diversity, are all tidied.}
}
\value{
None, Object of microtable itself cleaned up.
}
\description{
Trim the dataset to make OTUs and samples consistent across all files in the object.
}
\examples{
dataset$tidy_dataset(main_data = TRUE)
}
