% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_network.R
\name{trans_corr}
\alias{trans_corr}
\title{Create trans_corr object.}
\value{
res_cor_p in trans_corr object.
}
\description{
This class is a wrapper for a series of correlation calculating methods.
The functions in this class: \code{\link{replace_name}}.
}
\examples{
t1 <- trans_corr$new(dataset = dataset, cal_cor = "base", taxa_level = "OTU", filter_thres = 0.0001)
t1 <- trans_corr$new(dataset = dataset, cal_cor = "SparCC", taxa_level = "OTU", filter_thres = 0.0001)
t1 <- trans_corr$new(dataset = dataset, cal_cor = "WGCNA")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trans_corr$new()}}
\item \href{#method-replace_name}{\code{trans_corr$replace_name()}}
\item \href{#method-print}{\code{trans_corr$print()}}
\item \href{#method-clone}{\code{trans_corr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_corr$new(
  dataset = NULL,
  cor_method = c("pearson", "spearman", "kendall")[1],
  cal_cor = c("base", "WGCNA", "SparCC", "none")[1],
  taxa_level = "OTU",
  filter_thres = 0,
  nThreads = 1,
  SparCC_simu_num = 100,
  env_cols = NULL,
  add_data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} Class.}

\item{\code{cor_method}}{default "pearson"; one of c("pearson", "spearman", "kendall"); correlation algorithm.}

\item{\code{cal_cor}}{default "base"; one of  c("base", "WGCNA", "SparCC", "none"); correlation method; "none" represent do not calculate.}

\item{\code{taxa_level}}{default "OTU"; taxonomic rank.}

\item{\code{filter_thres}}{default 0; the relative abundance threshold.}

\item{\code{nThreads}}{default 1; the thread number used for "WGCNA" and SparCC.}

\item{\code{SparCC_simu_num}}{default 100; SparCC simulation number for bootstrap.}

\item{\code{env_cols}}{default NULL; number or name vector to select the physicochemical data in dataset$sample_table.}

\item{\code{add_data}}{default NULL; provide physicochemical table additionally.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_name"></a>}}
\if{latex}{\out{\hypertarget{method-replace_name}{}}}
\subsection{Method \code{replace_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_corr$replace_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_corr$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_corr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
