% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_norm.R
\name{trans_norm}
\alias{trans_norm}
\title{Feature abundance normalization/transformation.}
\description{
Feature abundance normalization/transformation for a microtable object or data.frame object.
}
\examples{

## ------------------------------------------------
## Method `trans_norm$new`
## ------------------------------------------------

library(microeco)
data(dataset)
t1 <- trans_norm$new(dataset = dataset)

## ------------------------------------------------
## Method `trans_norm$norm`
## ------------------------------------------------

newdataset <- t1$norm(method = "log")
newdataset <- t1$norm(method = "CLR")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-trans_norm-new}{\code{trans_norm$new()}}
\item \href{#method-trans_norm-norm}{\code{trans_norm$norm()}}
\item \href{#method-trans_norm-clone}{\code{trans_norm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_norm-new"></a>}}
\if{latex}{\out{\hypertarget{method-trans_norm-new}{}}}
\subsection{Method \code{new()}}{
Get a transposed abundance table if the input is microtable object. In the table, rows are samples, and columns are features. 
  This can make the further operations same with the traditional ecological methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_norm$new(dataset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the \code{\link{microtable}} object or \code{data.frame} object. 
If it is \code{data.frame} object, please make sure that rows are samples, and columns are features.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data_table, stored in the object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(microeco)
data(dataset)
t1 <- trans_norm$new(dataset = dataset)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_norm-norm"></a>}}
\if{latex}{\out{\hypertarget{method-trans_norm-norm}{}}}
\subsection{Method \code{norm()}}{
Normalization/transformation methods include CLR, CCS, TSS, TMM, AST and those based on the \code{\link{decostand}} function in vegan package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_norm$norm(
  method = NULL,
  MARGIN = NULL,
  logbase = exp(1),
  Cmin = NULL,
  pseudocount = 1,
  intersect.no = 10,
  ct.min = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{default NULL; See the following details and available options. \cr 
\cr 
Methods for normalization:
\itemize{
  \item \code{GMPR}: Geometric mean of pairwise ratios <doi: 10.7717/peerj.4600>. 
  \item \code{CLR}: Centered log-ratio normalization. 
  \item \code{CCS}: Cumulative sum scaling normalization based on the \code{metagenomeSeq} package.
  \item \code{TSS}: Total sum scaling, dividing counts by the sequencing depth.
  \item \code{TMM}: Trimmed mean of M-values method based on the \code{normLibSizes} function of \code{edgeR} package.
  \item \code{SRS}: scaling with ranked subsampling method based on the SRS package provided by Lukas Beule and Petr Karlovsky (2020) <DOI:10.7717/peerj.9593>.
}
Methods based on \code{\link{decostand}} function:
\itemize{
  \item \code{total}: divide by margin total (default MARGIN = 1, i.e. rows - samples).
  \item \code{max}: divide by margin maximum (default MARGIN = 2, i.e. columns - features).
  \item \code{normalize}:  make margin sum of squares equal to one (default MARGIN = 1).
  \item \code{range}: standardize values into range 0...1 (default MARGIN = 2). If all values are constant, they will be transformed to 0.
  \item \code{standardize}: scale x to zero mean and unit variance (default MARGIN = 2).
  \item \code{pa}: scale x to presence/absence scale (0/1).
  \item \code{log}: logarithmic transformation as suggested by Anderson et al. (2006): log_b (x) + 1 for x > 0, where b is the base of the logarithm; zeros are left as zeros. Higher bases give less weight to quantities and more to presences, and logbase = Inf gives the presence/absence scaling. Please note this is not log(x+1). Anderson et al. (2006) suggested this for their (strongly) modified Gower distance (implemented as method = "altGower" in vegdist), but the standardization can be used independently of distance indices.
}
Other methods for transformation:
\itemize{
  \item \code{AST}: Arc sine square root transformation.
}}

\item{\code{MARGIN}}{default NULL; 1 = samples, and 2 = features of abundance table; only available when method comes from \code{\link{decostand}} function.
If MARGIN is NULL, use the default value in decostand function.}

\item{\code{logbase}}{default exp(1); The logarithm base used in method = "log" or "CLR".}

\item{\code{Cmin}}{default NULL; see Cmin parameter in \code{SRS::SRS} function; Only available when \code{method = "SRS"}.
If not provided, use the minimum number across all the samples.}

\item{\code{pseudocount}}{default 1; add pseudocount for those features with 0 abundance when \code{method = "CLR"}.}

\item{\code{intersect.no}}{default 10; the intersecting taxa number between paired sample for \code{method = "GMPR"}.}

\item{\code{ct.min}}{default 1; the minimum number of counts required to calculate ratios for \code{method = "GMPR"}.}

\item{\code{...}}{parameters pass to \code{\link{decostand}} or \code{metagenomeSeq::cumNorm} when method = "CCS" or 
\code{edgeR::normLibSizes} when method = "TMM" or GMPR function (intersect.no and ct.min parameters) when method = "GMPR" (https://github.com/jchen1981/GMPR).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new microtable object or data.frame object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{newdataset <- t1$norm(method = "log")
newdataset <- t1$norm(method = "CLR")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_norm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-trans_norm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_norm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
