% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{trans_nullmodel}
\alias{trans_nullmodel}
\title{Create trans_nullmodel object.}
\value{
intermediate files in object.
}
\description{
This class is a wrapper for a series of null model related approaches.
The functions in this class: \code{\link{cal_mantel_corr}}, \code{\link{plot_mantel_corr}}, \code{\link{cal_betampd}}, \code{\link{cal_betamntd}},
\code{\link{cal_ses_betampd}}, \code{\link{cal_ses_betamntd}}, \code{\link{cal_rcbray}}, \code{\link{cal_process}}
}
\examples{
t1 <- trans_nullmodel$new(dataset, taxa_number = 1000, add_data = env_data, cpp = TRUE)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trans_nullmodel$new()}}
\item \href{#method-cal_mantel_corr}{\code{trans_nullmodel$cal_mantel_corr()}}
\item \href{#method-plot_mantel_corr}{\code{trans_nullmodel$plot_mantel_corr()}}
\item \href{#method-cal_betampd}{\code{trans_nullmodel$cal_betampd()}}
\item \href{#method-cal_betamntd}{\code{trans_nullmodel$cal_betamntd()}}
\item \href{#method-cal_ses_betampd}{\code{trans_nullmodel$cal_ses_betampd()}}
\item \href{#method-cal_ses_betamntd}{\code{trans_nullmodel$cal_ses_betamntd()}}
\item \href{#method-cal_rcbray}{\code{trans_nullmodel$cal_rcbray()}}
\item \href{#method-cal_process}{\code{trans_nullmodel$cal_process()}}
\item \href{#method-clone}{\code{trans_nullmodel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$new(
  dataset = NULL,
  filter_thres = 0,
  taxa_number = NULL,
  group = NULL,
  select_group = NULL,
  cpp = FALSE,
  env_cols = NULL,
  add_data = NULL,
  complete_na = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} Class.}

\item{\code{filter_thres}}{default 0; the relative abundance threshold.}

\item{\code{taxa_number}}{default NULL; how many taxa you want to use, if set, filter_thres parameter invalid.}

\item{\code{group}}{default NULL; which group column name in sample_table is selected.}

\item{\code{select_group}}{default NULL; the group name, used following the group to filter samples.}

\item{\code{cpp}}{default FALSE; invoke c++ to accelerate betaMPD calculation.}

\item{\code{env_cols}}{default NULL; number or name vector to select the physicochemical data in dataset$sample_table.}

\item{\code{add_data}}{default NULL; provide physicochemical table additionally.}

\item{\code{complete_na}}{default FALSE; whether fill the NA in physicochemical data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_mantel_corr"></a>}}
\if{latex}{\out{\hypertarget{method-cal_mantel_corr}{}}}
\subsection{Method \code{cal_mantel_corr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$cal_mantel_corr(
  use_env = NULL,
  break.pts = seq(0, 1, 0.02),
  cutoff = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_mantel_corr"></a>}}
\if{latex}{\out{\hypertarget{method-plot_mantel_corr}{}}}
\subsection{Method \code{plot_mantel_corr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$plot_mantel_corr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_betampd"></a>}}
\if{latex}{\out{\hypertarget{method-cal_betampd}{}}}
\subsection{Method \code{cal_betampd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$cal_betampd(abundance.weighted = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_betamntd"></a>}}
\if{latex}{\out{\hypertarget{method-cal_betamntd}{}}}
\subsection{Method \code{cal_betamntd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$cal_betamntd(
  abundance.weighted = FALSE,
  exclude.conspecifics = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_ses_betampd"></a>}}
\if{latex}{\out{\hypertarget{method-cal_ses_betampd}{}}}
\subsection{Method \code{cal_ses_betampd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$cal_ses_betampd(runs = 1000, abundance.weighted = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_ses_betamntd"></a>}}
\if{latex}{\out{\hypertarget{method-cal_ses_betamntd}{}}}
\subsection{Method \code{cal_ses_betamntd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$cal_ses_betamntd(
  runs = 1000,
  abundance.weighted = FALSE,
  exclude.conspecifics = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_rcbray"></a>}}
\if{latex}{\out{\hypertarget{method-cal_rcbray}{}}}
\subsection{Method \code{cal_rcbray()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$cal_rcbray(runs = 1000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_process"></a>}}
\if{latex}{\out{\hypertarget{method-cal_process}{}}}
\subsection{Method \code{cal_process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$cal_process(use_betamntd = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_nullmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
