% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_env.R
\name{trans_rda}
\alias{trans_rda}
\title{transform RDA result for the following plotting.}
\usage{
trans_rda(
  show_taxa = 10,
  adjust_arrow_length = FALSE,
  min_perc_env = 1,
  max_perc_env = 100,
  min_perc_tax = 1,
  max_perc_tax = 100
)
}
\arguments{
\item{show_taxa}{default 10; taxa number shown in the plot.}

\item{adjust_arrow_length}{default FALSE; whether adjust the arrow length to be clear}

\item{min_perc_env}{default 1; minimum value for env arrow, relatively.}

\item{max_perc_env}{default 100; maximum value for env arrow, relatively.}

\item{min_perc_tax}{default 1; minimum value for tax arrow, relatively.}

\item{max_perc_tax}{default 100; maximum value for tax arrow, relatively.}
}
\value{
res_rda_trans in object.
}
\description{
transform RDA result for the following plotting.
}
\examples{
dataset$trans_rda(adjust_arrow_length = TRUE, max_perc_env = 10)
}
