% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_venn.R
\name{trans_venn_com}
\alias{trans_venn_com}
\title{Transform venn result for the community analysis.}
\usage{
trans_venn_com(use_OTUs_frequency = TRUE)
}
\arguments{
\item{use_OTUs_frequency}{default TRUE; whether only use OTUs occurrence frequency, if FALSE, use abundance.}
}
\value{
a new \code{\link{microtable}} class.
}
\description{
Transform venn result for the community analysis.
}
\examples{
dataset$trans_venn_com()
}
