% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{clone}
\alias{clone}
\title{Copy an R6 class object completely}
\usage{
clone(x, deep = TRUE)
}
\arguments{
\item{x}{R6 class object}

\item{deep}{default TRUE; deep copy}
}
\value{
identical but unrelated R6 object.
}
\description{
Copy an R6 class object completely
}
\examples{
data("dataset")
clone(dataset)
}
