% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{The dataset in the microeco package}
\format{
An R6 class object
}
\usage{
data(dataset)
}
\description{
The dataset is structured with microtable class for the demonstration of examples and tutorials.
}
\details{
\itemize{
  \item sample_table: sample information table
  \item otu_table: species-community abundance table
  \item tax_table: taxonomic table
  \item phylo_tree: phylogenetic tree
  \item taxa_abund: taxa abundance list with several tables for Phylum...Genus
  \item alpha_diversity: alpha diversity table
  \item beta_diversity: list with several beta diversity distance matrix
}
}
\keyword{R6}
\keyword{object}
