% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fungi_func_FUNGuild}
\alias{fungi_func_FUNGuild}
\title{The FUNGuild database for fungi trait identification in the microeco package}
\usage{
data(fungi_func_FUNGuild)
}
\description{
The FUNGuild database for fungi trait identification in the microeco package
}
\keyword{data.frame}
