% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_nullmodel.R
\name{plot_mantel_corr}
\alias{plot_mantel_corr}
\title{Plot mantel correlogram.}
\usage{
plot_mantel_corr()
}
\value{
ggplot.
}
\description{
Plot mantel correlogram.
}
\examples{
dataset$plot_mantel_corr()
}
