% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prok_func_FAPROTAX}
\alias{prok_func_FAPROTAX}
\title{The modified FAPROTAX trait database in the microeco package}
\usage{
data(prok_func_FAPROTAX)
}
\description{
The modified FAPROTAX trait database in the microeco package
}
\keyword{data.frame}
