% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sparcc}
\alias{sparcc}
\title{sparcc wrapper}
\usage{
sparcc(data, iter = 20, inner_iter = 10, th = 0.1)
}
\arguments{
\item{data}{Community count data matrix}

\item{iter}{Number of iterations in the outer loop}

\item{inner_iter}{Number of iterations in the inner loop}

\item{th}{absolute value of correlations below this threshold are considered zero by the inner SparCC loop.}
}
\description{
An implementation of SparCC R code modified from <https://github.com/zdk123/SpiecEasi> based on Friedman et al. (2012) <doi:10.1371/journal.pcbi.1002687>.
}
\seealso{
\code{\link{sparccboot}}
}
