% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_table_16S}
\alias{taxonomy_table_16S}
\title{The taxonomic information of 16S dataset in the microeco package}
\usage{
data(taxonomy_table_16S)
}
\description{
The taxonomic information of 16S dataset in the microeco package
}
\keyword{data.frame}
