% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tidy_taxonomy}
\alias{tidy_taxonomy}
\title{Clear up the taxonomic table to make taxonomic assignments consistent.}
\format{
\code{\link{data.frame}} object.
}
\usage{
tidy_taxonomy(taxonomy_table)
}
\arguments{
\item{taxonomy_table}{a data.frame with taxonomic information.}
}
\value{
taxonomic table.
}
\description{
Clear up the taxonomic table to make taxonomic assignments consistent.
}
\examples{
data("taxonomy_table_16S")
tidy_taxonomy(taxonomy_table_16S)
}
