% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_beta.R
\name{trans_beta}
\alias{trans_beta}
\title{Create trans_beta Class to store and analyze beta-diveristy related statistics.}
\value{
res_ordination, group_distance_within, group_distance_between or res_manova in the object.
}
\description{
This class is a wrapper for a series of beta-diveristy related analysis.
The functions in this class include \code{\link{cal_group_distance}}, \code{\link{cal_manova}}, \code{\link{plot_ordination}},
\code{\link{plot_group_distance}}, \code{\link{plot_clustering}}
}
\examples{
trans_beta$new(dataset = dataset, ordination = "PCoA", measure = "bray", group = "Group", trans_between_group_distance = TRUE, cal_manova_paired = TRUE)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trans_beta$new()}}
\item \href{#method-cal_group_distance}{\code{trans_beta$cal_group_distance()}}
\item \href{#method-cal_manova}{\code{trans_beta$cal_manova()}}
\item \href{#method-plot_ordination}{\code{trans_beta$plot_ordination()}}
\item \href{#method-plot_group_distance}{\code{trans_beta$plot_group_distance()}}
\item \href{#method-plot_clustering}{\code{trans_beta$plot_clustering()}}
\item \href{#method-print}{\code{trans_beta$print()}}
\item \href{#method-clone}{\code{trans_beta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$new(
  dataset = NULL,
  ordination = NULL,
  measure = NULL,
  group = NULL,
  trans_otu = FALSE,
  ncomp = 3,
  scale.species = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} Class.}

\item{\code{ordination}}{default NULL; if provided, should be PCA, PCoA or NMDS.}

\item{\code{measure}}{default NULL; beta diversity index used for ordination, manova or group distance.}

\item{\code{group}}{default NULL; sample group used for manova or group distance.}

\item{\code{trans_otu}}{default FALSE; whether species abundance will be square transformed, used for PCA.}

\item{\code{scale.species}}{default FALSE; whether species loading in PCA will be scaled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_group_distance"></a>}}
\if{latex}{\out{\hypertarget{method-cal_group_distance}{}}}
\subsection{Method \code{cal_group_distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$cal_group_distance(within_group = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_manova"></a>}}
\if{latex}{\out{\hypertarget{method-cal_manova}{}}}
\subsection{Method \code{cal_manova()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$cal_manova(
  cal_manova_all = FALSE,
  cal_manova_paired = FALSE,
  cal_manova_set = NULL,
  permutations = 999
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_ordination"></a>}}
\if{latex}{\out{\hypertarget{method-plot_ordination}{}}}
\subsection{Method \code{plot_ordination()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$plot_ordination(
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  plot_color = NULL,
  plot_shape = NULL,
  plot_group_order = NULL,
  shape_values = c(16, 17, 7, 8, 15, 18, 11, 10, 12, 13, 9, 3, 4, 0, 1, 2, 14),
  plot_point_size = 3,
  plot_point_alpha = 0.9,
  plot_sample_label = NULL,
  plot_group_centroid = FALSE,
  plot_group = NULL,
  segment_alpha = 0.6,
  centroid_linetype = 3,
  plot_group_ellipse = FALSE,
  ellipse_level = 0.9,
  ellipse_alpha = 0.1,
  ellipse_type = "t"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_group_distance"></a>}}
\if{latex}{\out{\hypertarget{method-plot_group_distance}{}}}
\subsection{Method \code{plot_group_distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$plot_group_distance(
  plot_group_order = NULL,
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  distance_pair_stat = FALSE,
  pair_compare_filter = "",
  pair_compare_method = "wilcox.test",
  map_signif_level = TRUE,
  plot_distance_xtype = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_clustering"></a>}}
\if{latex}{\out{\hypertarget{method-plot_clustering}{}}}
\subsection{Method \code{plot_clustering()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$plot_clustering(
  use_colors = RColorBrewer::brewer.pal(8, "Dark2"),
  measure = NULL,
  group = NULL,
  replace_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measure}}{default NULL; beta diversity index used for ordination, manova or group distance.}

\item{\code{group}}{default NULL; sample group used for manova or group distance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_beta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
