% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_diff.R
\name{trans_diff}
\alias{trans_diff}
\title{Create trans_diff object for the difference tests on the taxonomic abundance and plotting.}
\value{
res_rf res_lefse res_abund or res_metastat in trans_diff object.
}
\description{
This class is a wrapper for a series of differential abundance test and indicator analysis methods.
The functions in this class include \code{\link{plot_diff_abund}}, \code{\link{plot_lefse_bar}}, \code{\link{plot_metastat}}
}
\examples{
trans_diff$new(dataset = dataset, method = "lefse", group = "Group")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trans_diff$new()}}
\item \href{#method-plot_diff_abund}{\code{trans_diff$plot_diff_abund()}}
\item \href{#method-plot_lefse_bar}{\code{trans_diff$plot_lefse_bar()}}
\item \href{#method-plot_metastat}{\code{trans_diff$plot_metastat()}}
\item \href{#method-print}{\code{trans_diff$print()}}
\item \href{#method-clone}{\code{trans_diff$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_diff$new(
  dataset = NULL,
  method = c("lefse", "rf", "metastat")[1],
  group = NULL,
  alpha = 0.05,
  lefse_subgroup = NULL,
  lefse_min_subsam = 10,
  lefse_norm = 1e+06,
  lefse_nresam = 0.6667,
  boots = 30,
  rf_taxa_level = "all",
  rf_ntree = 1000,
  metastat_taxa_level = "Genus",
  metastat_group_choose = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} Class.}

\item{\code{method}}{default "lefse"; one of c("lefse", "rf", "metastat").}

\item{\code{group}}{default NULL; sample group used for main comparision.}

\item{\code{alpha}}{default .05; significance threshold.}

\item{\code{lefse_subgroup}}{default NULL; sample sub group used in lefse.}

\item{\code{lefse_min_subsam}}{default 10; sample numbers required in the subgroup test.}

\item{\code{lefse_norm}}{default 1000000; scale value.}

\item{\code{lefse_nresam}}{default .6667; sample number ratio used in each bootstrap.}

\item{\code{boots}}{default 30; bootstrap test number for lefse or rf.}

\item{\code{rf_taxa_level}}{default "all"; use all taxonomic rank data, if want to test a specific rank, provide taxonomic rank name, such as "Genus".}

\item{\code{rf_ntree}}{default 1000; see ntree in \code{\link{randomForest}}.}

\item{\code{metastat_taxa_level}}{default "Genus"; taxonomic rank level used in metastat test.}

\item{\code{metastat_group_choose}}{default NULL; a vector used for selecting the required groups for testing.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_diff_abund"></a>}}
\if{latex}{\out{\hypertarget{method-plot_diff_abund}{}}}
\subsection{Method \code{plot_diff_abund()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_diff$plot_diff_abund(
  method = NULL,
  only_abund_plot = TRUE,
  use_number = 1:10,
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  plot1_bar_color = "grey50",
  plot2_sig_color = "red",
  plot2_sig_size = 1.2,
  axis_text_y = 10,
  simplify_names = TRUE,
  keep_prefix = TRUE,
  group_order = NULL,
  plot2_barwidth = 0.9,
  add_significance = TRUE,
  use_se = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{default "lefse"; one of c("lefse", "rf", "metastat").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_lefse_bar"></a>}}
\if{latex}{\out{\hypertarget{method-plot_lefse_bar}{}}}
\subsection{Method \code{plot_lefse_bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_diff$plot_lefse_bar(
  use_number = 1:10,
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  LDA_score = NULL,
  simplify_names = TRUE,
  keep_prefix = TRUE,
  group_order = NULL,
  axis_text_y = 12,
  plot_vertical = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_metastat"></a>}}
\if{latex}{\out{\hypertarget{method-plot_metastat}{}}}
\subsection{Method \code{plot_metastat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_diff$plot_metastat(
  use_number = 1:10,
  qvalue = 0.05,
  choose_group = 1,
  color_values = RColorBrewer::brewer.pal(8, "Dark2")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_diff$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_diff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
