% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_func.R
\name{trans_func}
\alias{trans_func}
\title{Create trans_func object for functional analysis.}
\value{
trans_func object with tax_table, otu_table, sample_table.
}
\description{
This class is a wrapper for a series of functional analysis for communities and species.
The functions in this class include \code{\link{cal_spe_func}}, \code{\link{cal_spe_func_perc}}, \code{\link{show_spe_func}},
\code{\link{plot_spe_func_perc}}, \code{\link{cal_tax4fun_func}}, \code{\link{cal_biogeo}}.
}
\examples{
t1 <- trans_diff$new(dataset = dataset)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trans_func$new()}}
\item \href{#method-cal_biogeo}{\code{trans_func$cal_biogeo()}}
\item \href{#method-cal_spe_func}{\code{trans_func$cal_spe_func()}}
\item \href{#method-cal_spe_func_perc}{\code{trans_func$cal_spe_func_perc()}}
\item \href{#method-show_spe_func}{\code{trans_func$show_spe_func()}}
\item \href{#method-plot_spe_func_perc}{\code{trans_func$plot_spe_func_perc()}}
\item \href{#method-cal_tax4fun_func}{\code{trans_func$cal_tax4fun_func()}}
\item \href{#method-print}{\code{trans_func$print()}}
\item \href{#method-clone}{\code{trans_func$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$new(dataset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} Class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_biogeo"></a>}}
\if{latex}{\out{\hypertarget{method-cal_biogeo}{}}}
\subsection{Method \code{cal_biogeo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$cal_biogeo(keep_tem = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_spe_func"></a>}}
\if{latex}{\out{\hypertarget{method-cal_spe_func}{}}}
\subsection{Method \code{cal_spe_func()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$cal_spe_func()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_spe_func_perc"></a>}}
\if{latex}{\out{\hypertarget{method-cal_spe_func_perc}{}}}
\subsection{Method \code{cal_spe_func_perc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$cal_spe_func_perc(use_community = TRUE, node_type_table = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_spe_func"></a>}}
\if{latex}{\out{\hypertarget{method-show_spe_func}{}}}
\subsection{Method \code{show_spe_func()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$show_spe_func(use_func = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_spe_func_perc"></a>}}
\if{latex}{\out{\hypertarget{method-plot_spe_func_perc}{}}}
\subsection{Method \code{plot_spe_func_perc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$plot_spe_func_perc(
  filter_func = NULL,
  group_list = NULL,
  group_list_default = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_tax4fun_func"></a>}}
\if{latex}{\out{\hypertarget{method-cal_tax4fun_func}{}}}
\subsection{Method \code{cal_tax4fun_func()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$cal_tax4fun_func(keep_tem = FALSE, folderReferenceData = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_func$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
